chore: move add-tool inside toolkit
