+++ src/app/toolkit/components/floatingButton.tsx
@@ -5,7 +5,7 @@ import Button from "@/ui/shared/Button";
 
 export default function FloatingButton() {
   return (
-    <Link href="/addTool">
+    <Link href="/toolkit/add-tool">
       <Button
         label="Add a Tool"
         className="fixed bottom-24 right-4 bg-twd-primary-purple text-white shadow-lg"
